CH-360 Moved test files to test/unit directory
