+++ provisioner/components/k8s/setup_playbook/final_setup/defaults/main.yml
