+++ provisioner/components/k8s/setup_playbook/final_setup/tasks/main.yml
@@ -11,3 +11,4 @@
     shell: |
       kubectl taint nodes --all node-role.kubernetes.io/master-
       kubectl taint nodes --all node-role.kubernetes.io/control-plane- 
+
