+++ provisioner/components/k8s/setup_playbook/final_setup/tests/test.yml
@@ -3,3 +3,4 @@
   remote_user: root
   roles:
     - final_setup
+
