+++ provisioner/test/unit/test_cluster_setup.py
@@ -17,25 +17,12 @@
 # For any questions about this software or licensing, please email
 # opensource@seagate.com or cortx-questions@seagate.com.
 
-from setup import setup
-from setup import createFile
-from teardown import teardown
-from validate import validate
-import os
-
-
-def test_folderCreation():
-    createFile()
-    x = os.system("ls /opt/halo/install_depot/ansible")
-    assert x==0, "File not created"
+from provisioner.components.k8s.cluster_setup import setup
+from provisioner.components.k8s.cluster_validate import validate
 
 
 def test_setup():
     setup()
     x = validate()
-    assert x!=0, "Software not installed"
+    assert x!=0, "Cluster setup failed"
 
-
-def test_teardown():
-    x = teardown()
-    assert x!=0, "Teardown Failed"
