+++ provisioner/test/unit/test_monitor.py
@@ -17,16 +17,16 @@
 # For any questions about this software or licensing, please email
 # opensource@seagate.com or cortx-questions@seagate.com.
 
-from setup import setup
-from setup import createFile
-from teardown import teardown
-from validate import validate
+from provisioner.components.monitor.setup import setup
+from provisioner.components.monitor.setup import createFile
+from provisioner.components.monitor.teardown import teardown
+from provisioner.components.monitor.validate import validate
 import os
 
 
 def test_folderCreation():
     createFile()
-    x = os.system("ls /opt/halo/install_depot/nodecli")
+    x = os.system("ls /opt/halo/install_depot/monitor")
     assert x==0, "File not created"
